feat(tok): v2 corpus.tok — u64 counts (fixes 14B-token overflow) + 16/24/32-bit packing
The legacy corpus.tok stored ntok/ndoc as u32 (max 4.29B) and every token as int32.
A 46 GB corpus is ~14B tokens -> the u32 count OVERFLOWED (~3.3x wrap), so the file
was structurally broken and training would silently read a truncated corpus. And a
16K-64K vocab needs only 16 bits, so int32 doubled the file.

New v2 format ("S0T2", tokmap.hpp): u64 ntok/ndoc + an explicit bits_per_token, with
the width chosen from the vocab -- byte-aligned 16/24/32 (2/3/4 bytes). vocab<=65536
lands at 16 bits, HALVING the token stream. A `flags` word reserves room for a future
sub-byte bit-packed layout (e.g. 20-bit) with no caller changes -- everything reads
through the width-agnostic TokView. Legacy S0TK/S0TD (int32) still load.

Module: TokWriter (streaming, packs the width, back-patches u64 counts), TokView
(operator[]/first/subspan/copy_to over any width), TokMap (mmaps + parses v2 + legacy,
materializes the doc index to u64). Configurator writes v2 via TokWriter; training
consumes TokView (the GPU step + CPU train_batch + eval/bench/autotemp materialize a
window to int, since the on-disk width may be < int).

Verified end to end: fineweb_smoke -> v2 corpus.tok is 624 MB (exactly half the int32
1.25 GB), size math header32 + ntok*2 + ndoc*8 to the byte, round-trip OK; regchk
train reads the v2 uint16 stream (6.75M tokens, sane loss). Builds CPU + CUDA.

+++ include/sub0/window.hpp
@@ -39,7 +39,7 @@ struct Window {
 // trainable tokens (every index used stays in [0, train_tok)). With an empty index the result is a
 // flat full-width window. Single-token documents are skipped (no input/target pair to learn from).
 inline Window sample_window(std::mt19937& rng, int T, std::size_t train_tok,
-                            std::span<const std::uint32_t> docs) {
+                            std::span<const std::uint64_t> docs) {
     const std::size_t Tsz  = static_cast<std::size_t>(T);
     const std::size_t full = Tsz + 1;   // a full window needs T inputs + the last shifted target
     if (docs.empty()) {
@@ -50,7 +50,7 @@ inline Window sample_window(std::mt19937& rng, int T, std::size_t train_tok,
     for (int tries = 0; tries < 8; ++tries) {
         const std::size_t pos = uni(rng);
         const std::size_t k = static_cast<std::size_t>(
-            std::upper_bound(docs.begin(), docs.end(), static_cast<std::uint32_t>(pos)) - docs.begin()) - 1;
+            std::upper_bound(docs.begin(), docs.end(), static_cast<std::uint64_t>(pos)) - docs.begin()) - 1;
         const std::size_t ds = docs[k];
         std::size_t de = (k + 1 < docs.size()) ? static_cast<std::size_t>(docs[k + 1]) : train_tok;
         if (de > train_tok) de = train_tok;                 // a doc straddling the train/val split
